+++ src/hooks/useResponsive.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import React, { useEffect, useState } from 'react';
 import { TABLET_WIDTH } from '@/constants';
-import { useEffect, useState } from 'react';
 
 /**
  * useResponsive
